Numerous changes related to the Network module.
Overhaul big data receive logic:
- Fix an issue where passing Constants.MaxBigPacket would discard the BigReceive object causing the next big packet to be the beginning of a big data stream.
- Fix an issue where the rented array length was not large enough.
- Better spannification.

Refactor big data send logic to make better use of ReadOnlySpan<byte>. Previously, it tried to mimic ASSS logic too closely.

Overhaul reliable receive logic (including RelThread). Update to handle when sequence number overflows and wraps back around. Added a helper class, CircularSequenceQueue<T> to have better separation of concerns.

Refactored handling of ACKs. Added some additional logic to round trip time deviation for ACKs of packets that were resent.

Updated elapsed time measurement logic to use the Stopwatch class. This means it should be a much higher resolution than before. This includes round trip time actually being accurate to the millisecond.

Fix thread synchronization issues accessing the ConnData. Added comments on each field, identifying how it's being synchronized.

Add settings to control the reliable receive window size (separate settings for player connection vs client connection).

Spannify IClientConnectionHandler and the BillingUdp module's implementation. Spannified in such a way that only the data sent is passed, not a larger buffer + length. Updated the BillingUdp module to not overrun the length for variable length packets.

Update the ?userdbadm command to print connection stats. This should help give an idea of the quality of the connection to the billing server. Move printing of identity to a separate subcommand.

Bandwidth limiting
- Changed to use Stopwatch for higher resolution.
- Changed the max send windows size (Net:SendAtOnce) in the BandwidthDefault module from 255 to 64. The large window size was causing issues when sending sized data. It appears Continuum's receive window size is less than 255. It still ACKs the packets outside of the window (and internally probably overwrites the previous data). I dont know what Continuum's receive window size is, but so far 64 has not caused sending of sized data to get stuck.

Updated Constants.MaxPacket to 520.
- This allows certain packets to be sent without needing to be fragmented using big packets (0x00 0x08 and 0x00 0x09).
  + A full static flag packet (0x22) containing 256 flags. The max length being: 1 + (256 * 2) = 513. Plus the 6 byte reliable header = 519. Tested it and it works.
  + This also allows a full brick packet (0x21) containing 32 bricks. The max length being: 1 + (32 * 16) = 513. Plus the 6 byte reliable header = 519. Tested and it it works.
- Updated the PeriodicReward module to limit the data it sends (0x23) to 513 bytes (the max that Continuum accepts), which with the 6 byte reliable header = 519.

+++ src/Core/Modules/LagData.cs
@@ -439,7 +439,7 @@ namespace SS.Core.Modules
             {
                 lock (lockObj)
                 {
-                    uint s, r;
+                    ulong s, r;
 
                     s = Packetloss.ServerPacketsSent;
                     r = Packetloss.ClientPacketsReceived;
